fix: Address 26 code review findings (CQ, UX, DOC, Security)
Code Quality:
- CQ1: Fix OOB write data[len]=0 in WebSocket handler -> String::concat(data,len)
- CQ2: glass_count uint16_t -> uint32_t (overflow after 65535)
- CQ3: setTaraWeight(uint16_t) -> float (implicit truncation fix)
- CQ4: Move OTA validation from blocking delay(10000) to non-blocking tickHousekeeping
- CQ5: MQTTServerData aggregate-init fixed with = {} (was missing server_topic)
- CQ6: ESPFS::readString() O(n^2) char-by-char -> file.readString()
- CQ7: Document template_response_buffer single-thread safety
- CQ8: readJsonConfig() double template read -> seededFromTemplate flag
- CQ9: first_run global -> static followUpActive in FOLLOW_UP block
- CQ10: Remove dead code (glass_full, log_event, getNTPDate, disconnect, restartESP)
- CQ11: Call validateAndFix() after WebSocket config changes
- CQ12: Remove redundant #pragma once (keep #ifndef guard)

Security:
- S3: Add bounds check in split() parser (key_value_idx < 3)

UX:
- UX1: Fix WLAN form parameter mismatch (ip_local->ip_address, ip_gw->gw, etc.)
- UX2: Add MQTT/API config backend handler + template variables
- UX3: Fix <dev> -> <div> tags and form nesting in setupwlan.html
- UX4: Version from template variable %VERSION% instead of hardcoded
- UX5: Redirect with ?saved=1 + green banner after config save
- UX7: beekeeping default "" to trigger first-run wizard
- UX8: Fix log levels (log_e -> log_i/log_d/log_w) in WebServerX + handleServo

Documentation:
- DOC4: Fix copyright STRATO AG -> Johannes G. Arlt in cooperative_loop.h, persistence.h
- DOC5: Rewrite README.md (purpose, hardware, setup, architecture, license)
- DOC6: Remove stale freertos_setup.cpp references in WebServerX.cpp
- DOC7: Add file headers to cooperative_loop.cpp, persistence.cpp
- DOC8: Add @brief Doxygen to all tick functions in cooperative_loop.h
- DOC9: Fix author "GitHub Copilot" -> "Johannes G. Arlt"

+++ lib/Glass/Glass.cpp
@@ -32,7 +32,9 @@ void Glass::reset() {
   cutoff_weight = HMConfig::instance().weight_filling;
 }
 
-void Glass::setTaraWeight(uint16_t tara_weight) { _glass_weight = tara_weight; }
+void Glass::setTaraWeight(float tara_weight) {
+  _glass_weight = static_cast<int16_t>(tara_weight);
+}
 
 /**
  * This function is allways called from main loop and handle most processing
